bot.py
import os
import imagehash
from PIL import Image
import telegram
from telegram import Message

# Telegram Bot Token (set in Render Environment)
TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN")

# Placeholder: Simulate Telegram image search
def search_image_in_telegram_groups(image_path):
    # Replace this with real Telegram API calls to fetch group images
    print(f"Searching for image: {image_path}")

    # Simulated comparison
    uploaded_hash = imagehash.average_hash(Image.open(image_path))

    # Simulate a list of previously seen image hashes (replace with real logic)
    known_images = {
        "https://t.me/samplegroup1/123": imagehash.hex_to_hash("ffeeddccbbaa9988"),
        "https://t.me/samplegroup2/456": imagehash.hex_to_hash("abcdabcdabcdabcd"),
    }

    matches = []
    for link, known_hash in known_images.items():
        if uploaded_hash - known_hash < 5:  # Hamming distance threshold
            matches.append(f"Found similar image: {link}")

    if not matches:
        matches.append("No matches found.")

    return matches