requirements.txt
Flask
python-telegram-bot==20.6
Pillow
ImageHash